+++ R/dwn_idbank_files.R
@@ -16,21 +16,58 @@ dwn_idbank_file = function(){
   option_proxy = Sys.getenv("INSEE_download_option_proxy")
   option_auth = Sys.getenv("INSEE_download_option_auth")
 
-  if (option_extra == ""){
-    dwn = utils::download.file(file_to_dwn, temp_file,
-                               mode = option_mode, quiet = TRUE)
+  # if (option_extra == ""){
+  #   dwn = utils::download.file(file_to_dwn, temp_file,
+  #                              mode = option_mode, quiet = TRUE)
+  # }else{
+  #   dwn = utils::download.file(file_to_dwn, temp_file,
+  #                              method = option_method,
+  #                              mode = option_mode,
+  #                              extra = option_extra,
+  #                              quiet = TRUE)
+  # }
+  httr::set_config(httr::config(ssl_verifypeer = FALSE))
+  # file_to_dwn = "https://www.insee.fr/en/statistiques/fichier/2868055/202209_correspondance_idbank_dimension.zip"
+
+  if(option_extra == ""){
+    response = try(httr::GET(file_to_dwn), silent = TRUE)
   }else{
-    dwn = utils::download.file(file_to_dwn, temp_file,
-                               method = option_method,
-                               mode = option_mode,
-                               extra = option_extra,
-                               quiet = TRUE)
+
+    proxy = httr::use_proxy(url = option_proxy,
+                            port = as.numeric(option_port),
+                            auth = option_auth)
+
+    response = httr::GET(url = file_to_dwn,
+                         config = proxy)
   }
 
+  # trigger error
+  if (response$status_code != 200){
+    trigger_error = "1" + "1"
+  }
 
-  uzp = utils::unzip(temp_file, exdir = insee_data_dir)
+  #create temporary directory for storing and unzipping file
+  td <- tempdir()
 
-  mapping_file = file.path(insee_data_dir, list.files(insee_data_dir, pattern = mapping_file_pattern)[1])
+  #open connection to write contents
+  zipF <- paste0(td, "/idbankfile.zip")
+
+  filecon <- file(zipF, "wb")
+
+  #write data contents to the temporary file
+  writeBin(response$content, filecon)
+
+  #close the connection
+  close(filecon)
+
+  uzp = utils::unzip(zipF, exdir = insee_data_dir)
+  potential_file = list.files(insee_data_dir, pattern = "correspondance_idbank_dimension")[1]
+
+  if (is.na(potential_file)){
+    potential_file = list.files(insee_data_dir)[1]
+  }
+
+  mapping_file = file.path(insee_data_dir, potential_file)
 
   mapping = utils::read.delim(mapping_file, sep = mapping_file_sep,
                               stringsAsFactors = F)
@@ -45,6 +82,7 @@ dwn_idbank_files = function(){
   if ("try-error" %in% class(data)){
 
     curr_year = as.numeric(substr(Sys.Date(), 1, 4))
+    curr_month = as.character(substr(Sys.Date(), 6, 7))
     last_year = curr_year - 1
 
     months_char = c()
@@ -56,9 +94,14 @@ dwn_idbank_files = function(){
       }
     }
 
-    dates_pattern_list = c(paste0(curr_year, months_char), paste0(last_year, months_char))
+    idbank_file_env = Sys.getenv("INSEE_idbank_dataset_path")
+    pattern_file_env = Sys.getenv("INSEE_idbank_dataset_file")
+
+    dates_pattern_list = c(paste0(curr_year, curr_month), paste0(curr_year, months_char), paste0(last_year, months_char))
     files_pattern = paste0(dates_pattern_list, "_correspondance_idbank_dimension")
+    files_pattern = c(pattern_file_env, files_pattern)
     files_dwn = paste0("https://www.insee.fr/en/statistiques/fichier/2868055/" , files_pattern, '.zip')
+    files_dwn = c(idbank_file_env, files_dwn)
 
     i = 1