+++ R/get_idbank_list.R
@@ -186,6 +186,13 @@ get_idbank_list = function(
         msg2 = "\nPackage's internal data has been used instead"
         msg3 = "\nPlease contact the package maintainer if this error persists"
         warning(sprintf("%s %s %s", msg1, msg2, msg3))
+        
+        msg = "\nIf a work computer is used, a proxy server may prevent this package from accessing the internet"
+        msg = paste0(msg, "\nIn this case, please ask your IT support team to provide you with the proxy server settings")
+        msg = paste0(msg, "\nThen, have a look at the following tutorial to use these settings:\n")
+        msg = paste0(msg, "https://cran.r-project.org/web/packages/insee/vignettes/insee.html")
+        
+        message(crayon::style(msg, "red"))
 
         idbank_list = tibble::as_tibble(idbank_list)