+++ R/get_insee.R
@@ -8,8 +8,6 @@
 #' Sys.setenv(INSEE_download_verbose = "FALSE").
 #' The use of cached data can be disabled with : Sys.setenv(INSEE_no_cache_use = "TRUE").
 #' All queries are printed in the console with this command: Sys.setenv(INSEE_print_query = "TRUE").
-#' The RapidXML C++ library is used as a backup thanks to the readsdmx package.
-#' It can be used instead of the internal parser with this command : Sys.setenv(INSEE_read_sdmx_fast = "TRUE")
 #' @param link SDMX query link
 #' @param step argument used only for internal package purposes to tweak download display
 #' @return a tibble containing the data
@@ -31,7 +29,6 @@ get_insee = function(link, step = "1/1"){
   insee_value_as_numeric = if(Sys.getenv("INSEE_value_as_numeric") == "TRUE"){TRUE}else{FALSE}
   insee_print_query = if(Sys.getenv("INSEE_print_query") == "TRUE"){TRUE}else{FALSE}
   insee_no_cache_use = if(Sys.getenv("INSEE_no_cache_use") == "TRUE"){TRUE}else{FALSE}
-  insee_read_sdmx_fast = if(Sys.getenv("INSEE_read_sdmx_fast") == "TRUE"){TRUE}else{FALSE}
 
   if(insee_download_verbose){
     if(insee_print_query == TRUE) {
@@ -46,45 +43,13 @@ get_insee = function(link, step = "1/1"){
 
   if((!file.exists(file_cache)) | insee_no_cache_use){
 
-    use_read_sdmx_fast_first = insee_read_sdmx_fast
-    use_backup_parser = TRUE
-
-    if(stringr::str_detect(link, "includeHistory")){
-      use_backup_parser = FALSE
-      insee_read_sdmx_fast = FALSE
-    }
-
-    if(link == Sys.getenv("INSEE_sdmx_link_dataflow")){
-      use_backup_parser = FALSE
-      insee_read_sdmx_fast = FALSE
-    }
-
-    # by default the internal parser is used, if it fails the readsdmx parser is used
-    if(!use_read_sdmx_fast_first){
-      data_final = read_sdmx_slow(link, step)
-
-      if(use_backup_parser){
-        if(is.null(data_final)){
-          data_final = read_sdmx_fast(link, step)
-        }
-      }
-
-    }else{
-      data_final = read_sdmx_fast(link, step)
-
-      if(use_backup_parser){
-        if(is.null(data_final)){
-          data_final = read_sdmx_slow(link, step)
-        }
-      }
-
-    }
+    data_final = read_sdmx_slow(link, step)
 
     if(!is.null(data_final)){
 
       s = try(saveRDS(data_final, file = file_cache), silent = TRUE)
 
-      if(class(s) != "try-error"){
+      if(!"try-error" %in% class(s)){
         if(insee_download_verbose){
           msg = sprintf("Data cached : %s\n", file_cache)
 
@@ -94,6 +59,10 @@ get_insee = function(link, step = "1/1"){
 
     }else{
       msg = "An error occurred"
+      msg = paste0(msg, "\n\nIf a work computer is used, a proxy server may prevent this package from accessing the internet")
+      msg = paste0(msg, "\nIn this case, please ask your IT support team to provide you with the proxy server settings")
+      msg = paste0(msg, "\nThen, have a look at the following tutorial to use these settings:\n")
+      msg = paste0(msg, "https://cran.r-project.org/web/packages/insee/vignettes/insee.html")
       message(crayon::style(msg, "red"))
     }
   }else{