+++ inst/doc/v6_pop_map-vignettes.R
@@ -104,19 +104,20 @@ embed_png("pop_map.png")
 #                                     pop_density >= 260 & pop_density < 410 ~ "[260, 410]",
 #                                     pop_density >= 410 & pop_density < 600 ~ "[410, 600]",
 #                                     pop_density >= 600 & pop_density < 1000 ~ "[600, 1000]",
+#                                     pop_density >= 1000 & pop_density < 5000 ~ "[1000, 5000]",
 #                                     pop_density >= 5000 & pop_density < 10000 ~ "[5000, 10000]",
 #                                     pop_density >= 20000 ~ ">= 20000"
 #    )) %>%
 #    mutate(`people per square kilometer` = factor(density_range,
-#                                  levels = c("< 40","[40, 50]", "[50, 70]","[70, 100]",
-#                                             "[100, 120]", "[120, 160]", "[160, 200]",
-#                                             "[200, 240]", "[240, 260]", "[260, 410]",
-#                                             "[410, 600]",  "[600, 1000]",
-#                                             "[5000, 10000]", ">= 20000")))
+#                                                  levels = c("< 40","[40, 50]", "[50, 70]","[70, 100]",
+#                                                             "[100, 120]", "[120, 160]", "[160, 200]",
+#                                                             "[200, 240]", "[240, 260]", "[260, 410]",
+#                                                             "[410, 600]",  "[600, 1000]", "[1000, 5000]",
+#                                                             "[5000, 10000]", ">= 20000")))
 #  
 #  ggplot(data = FranceMap_tidy_final_all,
-#                 aes(fill = `people per square kilometer`, x = long, y = lat, group = group) ,
-#                 size = 0, alpha = 0.9) +
+#         aes(fill = `people per square kilometer`, x = long, y = lat, group = group) ,
+#         size = 0, alpha = 0.9) +
 #    geom_polygon() +
 #    geom_path(colour = "white") +
 #    coord_map() +